Create classes folder and move Book, and Person within this.
