+++ classes/book/book_class.rb
