+++ classes/nameable_class.rb
