+++ classes/person/student_class.rb
