+++ classes/person/teacher_class.rb
